Replaced audio clips, edited DragDrop.cs so "correct" audio clip would play.
correct.mp3 is just the notification mp3 but trimmed so there's no empty space. incorrect.mp3 is just the family feud mp3 but trimmed for the same reason.
(Only for the TrainingScene_003_timerImplemented scene): For the button's script components, I assigned correct.mp3 to the "Correct" audio clip slots and incorrect.mp3 to the "Incorrect" audio clip slots.
In DragDrop.cs I replaced the method of hiding the button with a new function.
Also, I worked on this in a slightly older version of Unity.